pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Edantuti <edantuti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
psycopg-binary = "^3.0.16"
django-oauth-toolkit = "^2.1.0"
google-api-python-client = "^2.55.0"
opencv-python = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
